Convert MendixIdentifier to Long before using it in xpath query, to ensure compatibility with 7.23.x

+++ src/DeepLinkModule/javasource/deeplink/actions/StartDeeplinkJava.java
@@ -218,7 +218,7 @@ public class StartDeeplinkJava extends CustomJavaAction<java.lang.Boolean>
 							PendingLink.getType(),
 							PendingLink.MemberNames.PendingLink_DeepLink.toString(),
 							PendingLink.MemberNames.User.toString()))
-					.setVariable("mxId", deeplink.getMendixObject().getId())
+					.setVariable("mxId", deeplink.getMendixObject().getId().toLong())
 					.setVariable("username", userName)
 					.execute(sessionContext);
